+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `CocoDataset._alias_to_cat` looks up categories via the alias foreign key
 
 ### Changed
+* `id_to_idx` now behaves as a property.
 * Removed redundant items from `CategoryTree` serialization in `__json__` and `__getstate__`.
 * COG format is now defaulted over memmap frames backend
 * Using scalars for `load_sample`'s pad, will now broadcast to appropriate dimensions.